+++ src/common.ts
@@ -7,12 +7,14 @@ export {ProgressIndicatorOptions} from './modules/progressIndicator/services/pro
 export {ProgressIndicatorComponent} from './modules/progressIndicator/components/progressIndicator/progressIndicator.component';
 export {ProgressIndicatorModule} from './modules/progressIndicator/modules/progressIndicator.module';
 export {ProgressIndicatorService} from './modules/progressIndicator/services/progressIndicator.service';
+export {LocalProgressIndicatorName} from './modules/progressIndicator/misc/types';
 export {DebugDataModule} from './modules/debugData/modules/debugData.module';
 export {DebugDataComponent} from './modules/debugData/components/debugData/debugData.component';
 export {DebugDataEnabledService} from './modules/debugData/services/debugDataEnabled/debugDataEnabled.service';
 export {ClickOutsideDirective} from './modules/clickOutside/directives/clickOutside/clickOutside.directive';
 export {ClickOutsideModule} from './modules/clickOutside/modules/clickOutside.module';
 export * from './types/tokens';
+export * from './types/additionalInfo';
 export * from './utils';
 export {IsNaNPipe} from './pipes/isNaN/isNaN.pipe';
 export {IsPresentPipe} from './pipes/isPresent/isPresent.pipe';
@@ -24,7 +26,7 @@ export {CookiePermanentStorageService, PermanentStorage} from './services/perman
 export {MemoryTemporaryStorageService, TemporaryStorage} from './services/temporaryStorage';
 export {NoStringLocalizationService, StringLocalization} from './services/stringLocalization';
 export {Logger} from './services/logger';
-export {IgnoredInterceptorsService, IgnoredInterceptorId, HttpRequestIgnoredInterceptorId} from './services/ignoredInterceptors/ignoredInterceptors.service';
+export {IgnoredInterceptorsService, IgnoredInterceptorId} from './services/ignoredInterceptors/ignoredInterceptors.service';
 export {NgComponentOutletEx} from "./directives/ngComponentOutletEx/ngComponentOutletEx.directive";
 export {APP_STABLE, extractAppStableResolve} from "./utils";
-export {PROGRESS_INTERCEPTOR_PROVIDER, ProgressInterceptor} from './interceptors/progressInterceptor/progressInterceptor';
+export {PROGRESS_INTERCEPTOR_PROVIDER, ProgressInterceptor} from './modules/progressIndicator/interceptors/progressInterceptor';