+++ src/modules/progressIndicator/interceptors/progressInterceptor.ts
@@ -1,10 +1,12 @@
 import {Injectable, ClassProvider, Optional} from '@angular/core';
-import {HttpEvent, HttpInterceptor, HttpHandler, HttpEventType, HTTP_INTERCEPTORS} from '@angular/common/http';
+import {HttpEvent, HttpInterceptor, HttpHandler, HttpEventType, HTTP_INTERCEPTORS, HttpRequest} from '@angular/common/http';
 import {Observable} from 'rxjs';
 import {tap} from 'rxjs/operators';
 
-import {ProgressIndicatorService} from '../../modules/progressIndicator/services/progressIndicator.service';
-import {IgnoredInterceptorsService, HttpRequestIgnoredInterceptorId} from '../../services/ignoredInterceptors/ignoredInterceptors.service';
+import {ProgressIndicatorService} from '../services/progressIndicator.service';
+import {IgnoredInterceptorsService, IgnoredInterceptorId} from '../../../services/ignoredInterceptors/ignoredInterceptors.service';
+import {AdditionalInfo} from '../../../types/additionalInfo';
+import {LocalProgressIndicatorName} from '../misc/types';
 
 /**
  * ProgressInterceptor used for intercepting http requests and displaying progress indicatior
@@ -25,12 +27,12 @@ export class ProgressInterceptor implements HttpInterceptor
      * @param req - Request to be intercepted
      * @param next - Next middleware that can be called for next processing
      */
-    public intercept(req: HttpRequestIgnoredInterceptorId<any>, next: HttpHandler): Observable<HttpEvent<any>>
+    public intercept(req: HttpRequest<any> & AdditionalInfo<LocalProgressIndicatorName & IgnoredInterceptorId>, next: HttpHandler): Observable<HttpEvent<any>>
     {
         return next.handle(req)
             .pipe(tap(event =>
             {
-                if (this._ignoredInterceptorsService && this._ignoredInterceptorsService.isIgnored(ProgressInterceptor, req))
+                if (this._ignoredInterceptorsService && this._ignoredInterceptorsService.isIgnored(ProgressInterceptor, req.additionalInfo))
                 {
                     return;
                 }
@@ -38,14 +40,14 @@ export class ProgressInterceptor implements HttpInterceptor
                 //request started
                 if(event.type == HttpEventType.Sent)
                 {
-                    this._indicatorSvc.showProgress();
+                    this._indicatorSvc.showProgress(req.additionalInfo?.progressGroupName);
                 }
                 //response received
                 else if(event.type == HttpEventType.Response)
                 {
-                    this._indicatorSvc.hideProgress();
+                    this._indicatorSvc.hideProgress(req.additionalInfo?.progressGroupName);
                 }
-            }, () => this._indicatorSvc.hideProgress()));
+            }, () => this._indicatorSvc.hideProgress(req.additionalInfo?.progressGroupName)));
     }
 }